RAG Based Chatbot/requirements.txt
flask
numpy
sentence-transformers
faiss-cpu
langchain-community
langchain-text-splitters
pypdf
docx2txt
google-genai